+++ src/calculate_point_atmospheric_parameters.h
@@ -13,8 +13,8 @@ int calculate_point_atmospheric_parameters
 (
     Input_t *input,            /* I: input structure */
     REANALYSIS_POINTS *points, /* I: The coordinate points */
-    float albedo,              /* I: albedo */
-    float **results,           /* O: atmospheric parameter for modtarn run */
+    double albedo,             /* I: albedo */
+    double **results,          /* O: atmospheric parameter for modtarn run */
     bool verbose               /* I: value to indicate if intermediate
                                      messages will be printed */
 );
